Added error handling for files that do not exist
Added error handling for files that do not exist in a way that mimics grep on Linux. In its current form, it allows the use of directories, in which it will search all child files in that directory as well. Since grep on Linux does not dot this I will script around this in the future.