Renaming the 'PowerGrep' module.
Renaming the 'PowerGrep' module to 'PowerShellGrep', as there is currently an existing product out there called 'PowerGrep', which is a fat application used for advanced, regex based find and replace. This serves more as an interface/alias for the native PowerShell Select-String cmdlet, that allows it to be worked with like one would work with grep.